+++ lib/sync_queue.dart
@@ -1,6 +1,7 @@
 library;
 
 export 'src/models/sync_entity_ref.dart';
+export 'src/models/sync_entity_state.dart';
 export 'src/models/sync_failure.dart';
 export 'src/models/sync_operation.dart';
 export 'src/models/sync_record.dart';